# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Roadmap Accountability System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get user profile with stats
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        firstName:
          type: string
        currentPhase:
          type: string
        totalHoursLogged:
          type: number
        currentStreak:
          type: number
        longestStreak:
          type: number
        lastCheckIn:
          type: string
          nullable: true
        achievements:
          type: array
          items:
            type: string
      required: [id, username, currentPhase, totalHoursLogged, currentStreak, longestStreak, achievements]
    UpdateUserProfileBody:
      type: object
      properties:
        currentPhase:
          type: string
        firstName:
          type: string
...