# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Roadmap Accountability System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  Authorization: zod.string().optional(),
});

export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().nullish(),
      firstName: zod.string().nullish(),
      lastName: zod.string().nullish(),
      profileImageUrl: zod.string().nullish(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Start browser login
 */
export const BeginBrowserLoginQueryParams = zod.object({
  returnTo: zod.coerce.string().optional(),
});

/**
 * @summary Logout browser session
 */
export const LogoutBrowserSessionHeader = zod.object({
  Authorization: zod.string().optional(),
});

/**
 * @summary Exchange mobile OIDC code
 */

export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string().min(1),
  code_verifier: zod.string().min(1),
  redirect_uri: zod.string().min(1),
  state: zod.string().min(1),
  nonce: zod.string().min(1).optional(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Mobile logout
 */
export const LogoutMobileSessionHeader = zod.object({
  Authorization: zod.string().optional(),
});

export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get today's check-in if it exists
 */
export const GetTodayCheckinResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  date: zod.string(),
  hasUni: zod.boolean(),
  hasOffice: zod.boolean(),
  tasks: zod.string(),
  availableHours: zod.number(),
  pomodoroSessions: zod.number(),
  hoursLogged: zod.number(),
  aiSchedule: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Create or update today's check-in
 */
export const CreateCheckinBody = zod.object({
  hasUni: zod.boolean(),
  hasOffice: zod.boolean(),
  tasks: zod.string(),
  availableHours: zod.number(),
});

export const CreateCheckinResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  date: zod.string(),
  hasUni: zod.boolean(),
  hasOffice: zod.boolean(),
  tasks: zod.string(),
  availableHours: zod.number(),
  pomodoroSessions: zod.number(),
  hoursLogged: zod.number(),
  aiSchedule: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Get recent pomodoro sessions
 */
export const GetPomodoroSessionsQueryParams = zod.object({
  limit: zod.coerce.number().optional(),
});

export const GetPomodoroSessionsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  date: zod.string(),
  workDuration: zod.number(),
  breakDuration: zod.number(),
  completedAt: zod.string(),
});
export const GetPomodoroSessionsResponse = zod.array(
  GetPomodoroSessionsResponseItem,
);

/**
 * @summary Log a completed pomodoro session
 */
export const CreatePomodoroSessionBody = zod.object({
  workDuration: zod.number(),
  breakDuration: zod.number(),
});

export const CreatePomodoroSessionResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  date: zod.string(),
  workDuration: zod.number(),
  breakDuration: zod.number(),
  completedAt: zod.string(),
});

/**
 * @summary Get all episode completion records
 */
export const GetEpisodeProgressResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  episodeId: zod.string(),
  episodeTitle: zod.string(),
  phase: zod.string(),
  completedAt: zod.string(),
});
export const GetEpisodeProgressResponse = zod.array(
  GetEpisodeProgressResponseItem,
);

/**
 * @summary Mark an episode as completed
 */
export const LogEpisodeCompletionBody = zod.object({
  episodeId: zod.string(),
  episodeTitle: zod.string(),
  phase: zod.string(),
});

export const LogEpisodeCompletionResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  episodeId: zod.string(),
  episodeTitle: zod.string(),
  phase: zod.string(),
  completedAt: zod.string(),
});

/**
 * @summary Unmark an episode as completed
 */
export const RemoveEpisodeCompletionParams = zod.object({
  episodeId: zod.coerce.string(),
});

export const RemoveEpisodeCompletionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get hours logged (heatmap data)
 */
export const GetHoursLogQueryParams = zod.object({
  days: zod.coerce.number().optional(),
});

export const GetHoursLogResponseItem = zod.object({
  date: zod.string(),
  hours: zod.number(),
  sessions: zod.number(),
});
export const GetHoursLogResponse = zod.array(GetHoursLogResponseItem);

/**
 * @summary Get overall progress statistics
 */
export const GetProgressStatsResponse = zod.object({
  totalHours: zod.number(),
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  totalEpisodes: zod.number(),
  totalPomodoros: zod.number(),
  hoursThisWeek: zod.number(),
  hoursThisMonth: zod.number(),
  lastCheckIn: zod.string().nullish(),
});

/**
 * @summary Generate AI daily schedule
 */
export const GenerateScheduleBody = zod.object({
  hasUni: zod.boolean(),
  hasOffice: zod.boolean(),
  tasks: zod.string(),
  availableHours: zod.number(),
  currentPhase: zod.string(),
});

export const GenerateScheduleResponse = zod.object({
  schedule: zod.string(),
  tasks: zod.array(zod.string()),
  pomodoroRecommendation: zod.string(),
});

/**
 * @summary Generate AI weekly accountability report
 */
export const GenerateWeeklyReportResponse = zod.object({
  report: zod.string(),
  hoursLogged: zod.number(),
  episodesCompleted: zod.number(),
  nextWeekGoal: zod.string(),
  motivationalMessage: zod.string(),
});

/**
 * @summary Get user profile with stats
 */
export const GetUserProfileResponse = zod.object({
  id: zod.string(),
  username: zod.string(),
  firstName: zod.string().optional(),
  currentPhase: zod.string(),
  totalHoursLogged: zod.number(),
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  lastCheckIn: zod.string().nullish(),
  achievements: zod.array(zod.string()),
});

/**
 * @summary Update user profile
 */
export const UpdateUserProfileBody = zod.object({
  currentPhase: zod.string().optional(),
  firstName: zod.string().optional(),
});

export const UpdateUserProfileResponse = zod.object({
  id: zod.string(),
  username: zod.string(),
  firstName: zod.string().optional(),
  currentPhase: zod.string(),
  totalHoursLogged: zod.number(),
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  lastCheckIn: zod.string().nullish(),
  achievements: zod.array(zod.string()),
});