# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Roadmap Accountability System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: string;
  username: string;
  firstName?: string;
  currentPhase: string;
  totalHoursLogged: number;
  currentStreak: number;
  longestStreak: number;
  lastCheckIn?: string | null;
  achievements: string[];
}